First (partial!) pass at restructuring the X packages. Most importantly: actually automate the binary/package generation. Also, we're going to be putting xlibs+xserver in same package,   and putting xlibs+xserver+$APP (xterm) in packages as well. This means that they can be App Cat'd, reaching more users.
This works towards that, by parameterizing the xlib/xserver/keyboard
packages to build with respect to "XAPPID", indicating name of the
package we're building.  See Makefiles for more details.

Note that we hardcode these paths statically since AFAIK we *have* to do
so for the fonts in Xsdl (kdrive), so might as well hardcode the linking
paths and the like.  If we could do this all dynamically we'd be better
off in terms of not rebuilding such things.

ATM this will only prepare the binaries and needed files and version
control them--the actual package generation will still be done in
build.git as usual.

Very much a WIP, some design decisions might change as this progresses.